medication-py-cli/requirements.txt

click>=8.0.0
supabase>=1.0.3
python-dotenv>=1.0.0
